Affiche les erreurs de la même façon
